refactor: separate workspace index process contracts
